requirements.txt
selenium==4.12.0
mysql-connector-python==8.2.0
ftfy==6.1.1
tqdm==4.66.1
pandas==2.0.2
xlsxwriter==3.1.9
openpyxl==3.1.2
opencage==2.4.0
beautifulsoup4==4.12.2